Implementing dark mode. Dark mode for BlokNavbar
