+++ src/components/Blok/BlokMenu.jsx
@@ -7,10 +7,10 @@ const BlokMenu = ({ screenSize, activeMenu, setActiveMenu }) => {
 
     const { currentMenu, setCurrentMenu, appTheme, setAppTheme } = useBlokContext()
 
-    console.log(appTheme)
+
     return (
         <div className='h-screen'>
-            <nav className='fixed w-[250px] md:w-[300px] lg:relative p-[20px] md:p-[42px] border-solid border-r border-blok-color h-screen overflow-y-scroll z-[100] bg-white'>
+            <nav className='fixed w-[250px] md:w-[300px] lg:relative p-[20px] md:p-[42px] border-solid border-r border-blok-color h-screen overflow-y-scroll z-[100] bg-white dark:bg-dark-mode'>
                 <div className="logo pb-14 ps-3">
                     <img src={BlokLogo} className='w-[73px] h-[24px] block' alt="Blok logo" />
                 </div>
@@ -24,12 +24,12 @@ const BlokMenu = ({ screenSize, activeMenu, setActiveMenu }) => {
                 {
                     BlokMenuData[0].mainMenu.map((menus, index) => (
                         <div className="menu mb-10" key={index}>
-                            <p className="text-16 font-bold leading-5 text-blok-grey ps-3 pb-6" key={index}>{menus.menuGrouptitle}</p>
+                            <p className="text-16 font-bold leading-5 text-blok-grey dark:text-white ps-3 pb-6" key={index}>{menus.menuGrouptitle}</p>
                             <ul className="m-0 p-0 flex flex-col">
                                 {menus.menuGroupItems.map((items, index) => (
-                                    <li className="text-blok-grey text-16 my-1" key={index}>
+                                    <li className="text-blok-grey dark:text-white text-16 my-1" key={index}>
                                         <button
-                                            className={`flex ps-3 gap-5 items-center justify-start h-[45px] w-full hover:bg-slate-200 hover:rounded-8 hover:scale-[1.025] ${currentMenu === items.menuItem ? 'bg-blok-green font-bold text-blok-dark rounded-8' : ''}`}
+                                            className={`flex ps-3 gap-5 items-center justify-start h-[45px] w-full hover:bg-slate-200 hover:rounded-8 hover:scale-[1.025] dark:hover:text-blok-dark ${currentMenu === items.menuItem ? 'bg-blok-green font-bold text-blok-dark rounded-8' : ''}`}
                                             onClick={(e) => setCurrentMenu(items.menuItem)}
                                         >
                                             <img src={items.menuIcon} alt={items.menuIcon} className='' />
@@ -46,9 +46,9 @@ const BlokMenu = ({ screenSize, activeMenu, setActiveMenu }) => {
                     <ul>
                         {
                             BlokMenuData[0].supportMenu.map((menu, index) => (
-                                <li className="text-blok-grey text-16 my-1" key={index}>
+                                <li className="text-blok-grey dark:text-white text-16 my-1" key={index}>
                                     <button
-                                        className={`flex ps-3 gap-5 items-center justify-start h-[45px] w-full hover:bg-slate-200 hover:rounded-8 hover:scale-[1.025] ${currentMenu === menu.menuItem ? 'bg-blok-green font-bold text-blok-dark rounded-8' : ''}`}
+                                        className={`flex ps-3 gap-5 items-center justify-start h-[45px] w-full hover:bg-slate-200 hover:rounded-8 hover:scale-[1.025] dark:hover:text-blok-dark ${currentMenu === menu.menuItem ? 'bg-blok-green font-bold text-blok-dark rounded-8' : ''}`}
                                         onClick={(e) => setCurrentMenu(menu.menuItem)}
                                     >
                                         <img src={menu.menuIcon} alt={menu.menuIcon} className='hover:stroke-blok-dark-green' />
@@ -60,16 +60,16 @@ const BlokMenu = ({ screenSize, activeMenu, setActiveMenu }) => {
                     </ul>
                 </div>
 
-                <div className="border border-blok-dark px-2 p-1 flex justify-between rounded-8 gap-1 mb-5">
+                <div className="border border-blok-dark dark:border-white px-2 py-[6px] flex justify-between rounded-8 gap-1 mb-5">
                     {
                         BlokMenuData[0].themeSwitch.map((themes, index) => (
                             <button
                                 key={index}
-                                className={`${appTheme === themes.value ? 'bg-blok-dark-green text-white' : ''} text-14 flex gap-2 justify-start items-center py-2 px-4 rounded-8 text-blok-grey hover:bg-slate-200 hover:rounded-8 hover:scale-[1.025]`}
+                                className={`${appTheme === themes.value ? 'bg-blok-dark-green text-white' : ''} theme_svg_wrapper text-14 flex gap-2 justify-start items-center py-2 px-4 rounded-8 text-blok-grey hover:rounded-8 hover:scale-[1.025] hover:text-white hover:bg-blok-dark-green dark:text-white`}
                                 value={themes.value}
                                 onClick={(e) => setAppTheme(themes.value)}
                             >
-                                <img src={themes.themeIcon} alt="Lightmode"/>
+                                <object data={themes.themeIcon} type="image/svg+xml" alt="Lightmode"></object>
                                 <span>{themes.theme}</span>
                             </button>
                         ))
