+++ src/context/BlokContext.jsx
@@ -4,6 +4,7 @@ const BlokContext = createContext();
 
 export const BlokContextProvider = ({children}) => {
     const [appTheme, setAppTheme] = useState('light')
+    localStorage.setItem('appTheme', appTheme)
     const [currentMenu, setCurrentMenu] = useState('Dashboard')
     const [currentRange, setCurrentRange] = useState('3HRS')
