+++ tailwind.config.js
@@ -32,6 +32,7 @@ module.exports = {
 				'blok-red': '#FEEEF1',
                 'blok-green': '#EFF7F4',
                 'blok-component-grey': '#F6F6F6',
+                'dark-mode': '#070C0F',
 			},
 			borderWidth: {
 				1: "1px",
